Refactor polls/table.
