+++ frontend/src/components/polls/submit/SubmitPoll.tsx
@@ -12,6 +12,21 @@ import { ToastContext } from "../../utils/Toast";
 import { queryCache, useMutation } from "react-query";
 import { AxiosError } from "axios";
 
+const [questionPlaceholder, choice0Placeholder, choice1Placeholder] = getRandom(
+    [
+        [
+            "Le portail…",
+            "C’était pas mieux avant.",
+            "C’est bien mieux maintenant.",
+        ],
+        ["La piche…", "C’était mieux avant.", "C’est moins bien maintenant."],
+        ["Le BDE…", "C’était mieux avant.", "C’est moins bien maintenant."],
+        ["Ton premier choix ?", "L’X", "Ulm"],
+        ["Le plus claqué ?", "L’Octo", "La biéro"],
+        ["Les plus sharks ?", "(La) JuMP", "Le Trium"],
+    ]
+);
+
 export const SubmitPoll = () => {
     const { sendSuccessToast, sendErrorToast } = useContext(ToastContext);
     const [create] = useMutation(api.polls.create, {
@@ -30,23 +45,6 @@ export const SubmitPoll = () => {
             );
         },
     });
-
-    const [
-        questionPlaceholder,
-        choice0Placeholder,
-        choice1Placeholder,
-    ] = getRandom([
-        [
-            "Le portail…",
-            "C’était pas mieux avant.",
-            "C’est bien mieux maintenant.",
-        ],
-        ["La piche…", "C’était mieux avant.", "C’est moins bien maintenant."],
-        ["Le BDE…", "C’était mieux avant.", "C’est moins bien maintenant."],
-        ["Ton premier choix ?", "L’X", "Ulm"],
-        ["Le plus claqué ?", "L’Octo", "La biéro"],
-        ["Les plus sharks ?", "(La) JuMP", "Le Trium"],
-    ]);
 
     const onSubmit = (values, { resetForm, setSubmitting }) => {
         let data = {
