+++ frontend/src/components/polls/table/polls-table.css
@@ -1,7 +1,3 @@
-.polls-table {
-    background-color: white;
-}
-
 .pollChoice {
     text-indent: 5%;
 }
