+++ frontend/src/services/api/polls.ts
@@ -1,12 +1,6 @@
 import { Poll } from "../../models/polls";
 import { AxiosResponse } from "axios";
 import { apiService, PaginatedResponse, unwrap } from "../apiService";
-import { sortingToApiParameter } from "../../components/utils/table/sorting";
-import {
-    PollStateFilter,
-    pollStateFilterToApiParameter,
-} from "../../components/polls/table/PollsTableFilter";
-import { joinNonEmpty } from "../../utils/parameter";
 import { toUrlParams } from "../../utils/urlParam";
 
 /**
@@ -24,17 +18,6 @@ const parseDates = (response: AxiosResponse<PaginatedResponse<Poll[]>>) => {
 
     return response;
 };
-
-const listGeneric = (parameters: string = "") => (page = 1) =>
-    unwrap<PaginatedResponse<Poll[]>>(
-        apiService
-            .get<PaginatedResponse<Poll[]>>(
-                `/polls/?page=${page}${
-                    parameters === undefined ? "" : "&"
-                }${parameters}`
-            )
-            .then(parseDates)
-    );
 
 export const polls = {
     list: (
@@ -63,38 +46,6 @@ export const polls = {
                 )
                 .then(parseDates)
         ),
-    /**
-     * List all the polls.
-     * @param userFilter if different from "", only list the polls created by `userFilter`; otherwise, list all the polls.
-     * @param stateFilter given to `pollStateFilterToApiParameter` to create a `state` parameter.
-     * @param sorting given to `sortingToApiParameter` to create an `ordering` parameter.
-     * @param page the page to load.
-     */
-    listAll: (
-        {
-            userFilter,
-            stateFilter,
-            sorting,
-        }: { userFilter: string; stateFilter: PollStateFilter; sorting: any },
-        page = 1
-    ) =>
-        listGeneric(
-            // Put all the parameters into an array, then join them with `&`.
-            joinNonEmpty(
-                [
-                    "page_size=10",
-                    userFilter !== "" ? `user=${userFilter}` : "",
-                    pollStateFilterToApiParameter(stateFilter, "state"),
-                    sortingToApiParameter(sorting, {
-                        question: "question",
-                        publicationDate: "publication_date",
-                        user: "user__pk",
-                        state: "state",
-                    }),
-                ],
-                "&"
-            )
-        )(page),
     get: (pollId) => unwrap<Poll>(apiService.get(`/polls/${pollId}/`)),
     create: ({
         data,
